Continue to minimize the landing page
In an effort to avoid the site visitor being navigated away from our page, and into some new (potentially overwhelming) site like github, I made a few big changes to our readme.

Big one: I changed the patchwork install page from github to the one that Mikey made for our onboarding link generator.  This way a new visitor just needs to click a single button and they are ready to go.

I also removed the opening link to our "applications" page, as I think this page is more confusing than helpful for the first-time visitor.

I also changed the wording for the "about this handbook" so that a new visitor can find the FAQ faster, and not be confused by a list of other sites they should check out and also consolidate.  I feel that section would be better served in the FAQ about helping.